Add v1.6 of 'Day 1 > Puzzle 1' java solution
In order to maintain our cursor within [0, 99] we must calculate a positive remainder, not only the remainder. Previously, the essence of the algorithm was exactly the same, the difference was that we were rotating across 199 values [-99, 99] because we weren't calculating the positive remainder. You could think on v1.0 as a circle with 200 numbers: where the left side would have negative numbers, the right side positive ones and both at the top and bottom we would have zero. On the other side, v1.6 is a more concise circle with only 100 numbers: where we would only have values within [0, 99] and a single zero.